+++ frontend/test/pytest/test_custom_devices.py
@@ -21,6 +21,7 @@ import pytest
 from catalyst import measure, qjit
 from catalyst.compiler import get_lib_path
 from catalyst.utils.exceptions import CompileError
+from catalyst.utils.runtime import device_get_toml_config, extract_backend_info
 
 # These have to match the ones in the configuration file.
 OPERATIONS = [
@@ -81,6 +82,32 @@ OPERATIONS = [
     "Adjoint(ISWAP)",
     "Adjoint(SISWAP)",
     "MultiControlledX",
+    "C(PauliY)",
+    "C(RY)",
+    "C(PauliX)",
+    "C(RX)",
+    "C(IsingXX)",
+    "C(Hadamard)",
+    "C(SWAP)",
+    "C(IsingYY)",
+    "C(S)",
+    "C(MultiRZ)",
+    "C(PhaseShift)",
+    "C(T)",
+    "C(IsingXY)",
+    "C(PauliZ)",
+    "C(Rot)",
+    "C(IsingZZ)",
+    "C(RZ)",
+    "C(SingleExcitationPlus)",
+    "C(GlobalPhase)",
+    "C(DoubleExcitationPlus)",
+    "C(SingleExcitationMinus)",
+    "C(DoubleExcitation)",
+    "GlobalPhase",
+    "C(SingleExcitation)",
+    "C(DoubleExcitationMinus)",
+    "BlockEncode",
 ]
 OBSERVABLES = [
     "PauliX",
@@ -91,18 +118,21 @@ OBSERVABLES = [
     "Identity",
     "Projector",
     "Hamiltonian",
+    "SparseHamiltonian",
     "Sum",
     "SProd",
     "Prod",
     "Exp",
 ]
 
+RUNTIME_LIB_PATH = get_lib_path("runtime", "RUNTIME_LIB_DIR")
+
 
 @pytest.mark.skipif(
-    not pathlib.Path(get_lib_path("runtime", "RUNTIME_LIB_DIR") + "/libdummy_device.so").is_file(),
+    not pathlib.Path(RUNTIME_LIB_PATH + "/libdummy_device.so").is_file(),
     reason="lib_dummydevice.so was not found.",
 )
-def test_custom_device():
+def test_custom_device_load():
     """Test that custom device can run using Catalyst."""
 
     class DummyDevice(qml.QubitDevice):
@@ -114,12 +144,12 @@ def test_custom_device():
         version = "0.0.1"
         author = "Dummy"
 
-        # Doesn't matter as at the moment it is dictated by QJITDevice
         operations = OPERATIONS
         observables = OBSERVABLES
 
         def __init__(self, shots=None, wires=None):
             super().__init__(wires=wires, shots=shots)
+            self._option1 = 42
 
         def apply(self, operations, **kwargs):
             """Unused"""
@@ -133,8 +163,14 @@ def test_custom_device():
 
             return "DummyDevice", get_lib_path("runtime", "RUNTIME_LIB_DIR") + "/libdummy_device.so"
 
+    device = DummyDevice(wires=1)
+    config = device_get_toml_config(device)
+    backend_info = extract_backend_info(device, config)
+    assert backend_info.kwargs["option1"] == 42
+    assert "option2" not in backend_info.kwargs
+
     @qjit
-    @qml.qnode(DummyDevice(wires=1))
+    @qml.qnode(device)
     def f():
         """This function would normally return False.
         However, DummyDevice as defined in libdummy_device.so
@@ -176,6 +212,38 @@ def test_custom_device_bad_directory():
 
     with pytest.raises(
         CompileError, match="Device at this-file-does-not-exist.so cannot be found!"
+    ):
+
+        @qjit
+        @qml.qnode(DummyDevice(wires=1))
+        def f():
+            return measure(0)
+
+
+def test_custom_device_no_c_interface():
+    """Test that custom device error."""
+
+    class DummyDevice(qml.QubitDevice):
+        """Dummy Device"""
+
+        name = "Dummy Device"
+        short_name = "dummy.device"
+        pennylane_requires = "0.33.0"
+        version = "0.0.1"
+        author = "Dummy"
+
+        operations = OPERATIONS
+        observables = OBSERVABLES
+
+        def __init__(self, shots=None, wires=None):
+            super().__init__(wires=wires, shots=shots)
+
+        def apply(self, operations, **kwargs):
+            """Unused."""
+            raise RuntimeError("Dummy device")
+
+    with pytest.raises(
+        CompileError, match="The dummy.device device does not provide C interface for compilation."
     ):
 
         @qjit